+++ calculations.h
@@ -5,7 +5,7 @@ int calculateEstimatedWaitingTime(int patientCount, int specialtyIDs[], int ward
 float calculateEmergencySurcharge(int triageLevel, int specialtyID);
 int calculateWardStayCost(int wardID, int daysAdmitted);
 float calculateTotalBill(int baseConsultationFee, float surcharge, int wardStayCost);
-float calculateAgeSubsidyDiscount(int age, int totalBillValue);
+float calculateAgeSubsidyDiscount(int age, float totalBillValue);
 int findFreeBed(int wardID, int bedOccupancy[4][20], int totalBedCapacities[]);
 
 #endif // CALCULATIONS_H_INCLUDED
